+++ test/unit/data_test.exs
@@ -163,7 +163,7 @@ defmodule RDF.DataTest do
                    RDF.Term.value(RDF.iri(EX.O2))
                  ],
                  RDF.Term.value(EX.p2()) => [RDF.Term.value(RDF.iri(EX.O3))],
-                 RDF.Term.value(EX.p3()) => ["_:foo", "bar"]
+                 RDF.Term.value(EX.p3()) => ["bar", "_:foo"]
                }
     end
 
@@ -335,7 +335,7 @@ defmodule RDF.DataTest do
                      RDF.Term.value(RDF.iri(EX.O2))
                    ],
                    RDF.Term.value(EX.p2()) => [RDF.Term.value(RDF.iri(EX.O3))],
-                   RDF.Term.value(EX.p3()) => ["_:foo", "bar"]
+                   RDF.Term.value(EX.p3()) => ["bar", "_:foo"]
                  },
                  RDF.Term.value(RDF.iri(EX.S2)) => %{
                    RDF.Term.value(EX.p2()) => [
@@ -507,7 +507,7 @@ defmodule RDF.DataTest do
                        RDF.Term.value(RDF.iri(EX.O2))
                      ],
                      RDF.Term.value(EX.p2()) => [RDF.Term.value(RDF.iri(EX.O3))],
-                     RDF.Term.value(EX.p3()) => ["_:foo", "bar"]
+                     RDF.Term.value(EX.p3()) => ["bar", "_:foo"]
                    },
                    RDF.Term.value(RDF.iri(EX.S2)) => %{
                      RDF.Term.value(EX.p2()) => [
@@ -523,7 +523,7 @@ defmodule RDF.DataTest do
                        RDF.Term.value(RDF.iri(EX.O2))
                      ],
                      RDF.Term.value(EX.p2()) => [RDF.Term.value(RDF.iri(EX.O3))],
-                     RDF.Term.value(EX.p3()) => ["_:foo", "bar", RDF.Term.value(RDF.iri(EX.O5))]
+                     RDF.Term.value(EX.p3()) => ["bar", "_:foo", RDF.Term.value(RDF.iri(EX.O5))]
                    },
                    RDF.Term.value(RDF.iri(EX.S3)) => %{
                      RDF.Term.value(EX.p3()) => [
@@ -555,7 +555,7 @@ defmodule RDF.DataTest do
                        RDF.Term.value(RDF.iri(EX.O2))
                      ],
                      p2: [RDF.Term.value(RDF.iri(EX.O3))],
-                     p3: ["_:foo", "bar"]
+                     p3: ["bar", "_:foo"]
                    },
                    RDF.Term.value(RDF.iri(EX.S2)) => %{
                      p2: [
@@ -571,7 +571,7 @@ defmodule RDF.DataTest do
                        RDF.Term.value(RDF.iri(EX.O2))
                      ],
                      p2: [RDF.Term.value(RDF.iri(EX.O3))],
-                     p3: ["_:foo", "bar", RDF.Term.value(RDF.iri(EX.O5))]
+                     p3: ["bar", "_:foo", RDF.Term.value(RDF.iri(EX.O5))]
                    },
                    RDF.Term.value(RDF.iri(EX.S3)) => %{
                      p3: [